Refactor: EmptyPlant 컴포넌트 분리
